CVchecker/backend/requirements.txt
flask==2.0.1
werkzeug==2.0.1
PyPDF2==3.0.1
python-docx==0.8.11
nltk==3.6.2